Clean up error throwing stubbing tests
