+++ Tests/MockingbirdTests/Framework/StubbingTests.swift
@@ -20,21 +20,9 @@ class StubbingTests: BaseTestCase {
   var childProtocol: ChildProtocolMock!
   var childProtocolInstance: ChildProtocol { return childProtocol }
   
-  var throwingProtocol: ThrowingProtocolMock!
-  var throwingProtocolInstance: ThrowingProtocol { return throwingProtocol }
-  
-  var rethrowingProtocol: RethrowingProtocolMock!
-  var rethrowingProtocolInstance: RethrowingProtocol { return rethrowingProtocol }
-  
-  var inoutProtocol: InoutProtocolMock!
-  var inoutProtocolInstance: InoutProtocol { return inoutProtocol }
-  
   override func setUp() {
     child = mock(Child.self)
     childProtocol = mock(ChildProtocol.self)
-    throwingProtocol = mock(ThrowingProtocol.self)
-    rethrowingProtocol = mock(RethrowingProtocol.self)
-    inoutProtocol = mock(InoutProtocol.self)
   }
   
   func testStubTrivialMethod_onClassMock_implicitlyStubbed() {
@@ -359,117 +347,6 @@ class StubbingTests: BaseTestCase {
     given(childProtocol.getChildInstanceVariable()) ~> {true}
     XCTAssertTrue(childProtocolInstance.childInstanceVariable)
     verify(childProtocol.getChildInstanceVariable()).wasCalled()
-  }
-  
-  // MARK: Throwing errors
-  
-  func testStubThrowingMethod_returnsValue() {
-    given(throwingProtocol.throwingMethod()) ~> true
-    XCTAssertTrue(try throwingProtocolInstance.throwingMethod())
-    verify(throwingProtocol.throwingMethod()).returning(Bool.self).wasCalled()
-  }
-  func testStubThrowingMethod_throwsError() {
-    given(throwingProtocol.throwingMethod()) ~> { () throws -> Bool in throw FakeError() }
-    XCTAssertThrowsError(try throwingProtocolInstance.throwingMethod() as Bool)
-    verify(throwingProtocol.throwingMethod()).returning(Bool.self).wasCalled()
-  }
-  func testStubParameterizedThrowingMethod_throwsError() {
-    given(throwingProtocol.throwingMethod(block: any())) ~> { _ in throw FakeError() }
-    XCTAssertThrowsError(try throwingProtocolInstance.throwingMethod(block: { true }))
-    verify(throwingProtocol.throwingMethod(block: any())).wasCalled()
-  }
-  func testStubParameterizedThrowingMethod_implicitlyRethrowsError() {
-    given(throwingProtocol.throwingMethod(block: any())) ~> { _ = try $0() }
-    XCTAssertThrowsError(try throwingProtocolInstance.throwingMethod(block: { throw FakeError() }))
-    verify(throwingProtocol.throwingMethod(block: any())).wasCalled()
-  }
-  
-  func testStubThrowingMethod_returnsValue_explicitSyntax() {
-    given(throwingProtocol.throwingMethod()).willReturn(true)
-    XCTAssertTrue(try throwingProtocolInstance.throwingMethod())
-    verify(throwingProtocol.throwingMethod()).returning(Bool.self).wasCalled()
-  }
-  func testStubThrowingMethod_throwsError_explicitSyntax() {
-    given(throwingProtocol.throwingMethod()).returning(Bool.self).willThrow(FakeError())
-    XCTAssertThrowsError(try throwingProtocolInstance.throwingMethod() as Bool)
-    verify(throwingProtocol.throwingMethod()).returning(Bool.self).wasCalled()
-  }
-  func testStubParameterizedThrowingMethod_throwsError_explicitSyntax() {
-    given(throwingProtocol.throwingMethod(block: any())).willThrow(FakeError())
-    XCTAssertThrowsError(try throwingProtocolInstance.throwingMethod(block: { true }))
-    verify(throwingProtocol.throwingMethod(block: any())).wasCalled()
-  }
-  func testStubParameterizedThrowingMethod_implicitlyRethrowsError_explicitSyntax() {
-    given(throwingProtocol.throwingMethod(block: any())).will { _ = try $0() }
-    XCTAssertThrowsError(try throwingProtocolInstance.throwingMethod(block: { throw FakeError() }))
-    verify(throwingProtocol.throwingMethod(block: any())).wasCalled()
-  }
-  
-  func testStubRethrowingReturningMethod_returnsValue() {
-    given(rethrowingProtocol.rethrowingMethod(block: any())) ~> true
-    XCTAssertTrue(try rethrowingProtocolInstance.rethrowingMethod(block: { throw FakeError() }))
-    verify(rethrowingProtocol.rethrowingMethod(block: any())).returning(Bool.self).wasCalled()
-  }
-  func testStubRethrowingReturningMethod_returnsValueFromBlock() {
-    given(rethrowingProtocol.rethrowingMethod(block: any())) ~> { return try $0() }
-    XCTAssertTrue(rethrowingProtocolInstance.rethrowingMethod(block: { return true }))
-    verify(rethrowingProtocol.rethrowingMethod(block: any())).returning(Bool.self).wasCalled()
-  }
-  func testStubRethrowingReturningMethod_rethrowsError() {
-    given(rethrowingProtocol.rethrowingMethod(block: any())) ~> { return try $0() }
-    XCTAssertThrowsError(try rethrowingProtocolInstance.rethrowingMethod(block: {
-      throw FakeError()
-    }) as Bool)
-    verify(rethrowingProtocol.rethrowingMethod(block: any())).returning(Bool.self).wasCalled()
-  }
-  func testStubRethrowingNonReturningMethod_rethrowsError() {
-    given(rethrowingProtocol.rethrowingMethod(block: any())) ~> { _ = try $0() }
-    XCTAssertThrowsError(try rethrowingProtocolInstance.rethrowingMethod(block: {
-      throw FakeError()
-    }) as Void)
-    verify(rethrowingProtocol.rethrowingMethod(block: any())).returning(Void.self).wasCalled()
-  }
-  
-  func testStubRethrowingReturningMethod_returnsValue_explicitSyntax() {
-    given(rethrowingProtocol.rethrowingMethod(block: any())).willReturn(true)
-    XCTAssertTrue(try rethrowingProtocolInstance.rethrowingMethod(block: { throw FakeError() }))
-    verify(rethrowingProtocol.rethrowingMethod(block: any())).returning(Bool.self).wasCalled()
-  }
-  func testStubRethrowingReturningMethod_returnsValueFromBlock_explicitSyntax() {
-    given(rethrowingProtocol.rethrowingMethod(block: any())).will { return try $0() }
-    XCTAssertTrue(rethrowingProtocolInstance.rethrowingMethod(block: { return true }))
-    verify(rethrowingProtocol.rethrowingMethod(block: any())).returning(Bool.self).wasCalled()
-  }
-  func testStubRethrowingReturningMethod_rethrowsError_explicitSyntax() {
-    given(rethrowingProtocol.rethrowingMethod(block: any())).will { return try $0() }
-    XCTAssertThrowsError(try rethrowingProtocolInstance.rethrowingMethod(block: {
-      throw FakeError()
-    }) as Bool)
-    verify(rethrowingProtocol.rethrowingMethod(block: any())).returning(Bool.self).wasCalled()
-  }
-  func testStubRethrowingNonReturningMethod_rethrowsError_explicitSyntax() {
-    given(rethrowingProtocol.rethrowingMethod(block: any())).will { _ = try $0() }
-    XCTAssertThrowsError(try rethrowingProtocolInstance.rethrowingMethod(block: {
-      throw FakeError()
-    }) as Void)
-    verify(rethrowingProtocol.rethrowingMethod(block: any())).returning(Void.self).wasCalled()
-  }
-  
-  // MARK: Inout parameters
-  
-  func testInoutParameter_doesNotMutateString() {
-    given(inoutProtocol.parameterizedMethod(object: any())) ~> { _ in }
-    var valueType = "foo bar"
-    inoutProtocolInstance.parameterizedMethod(object: &valueType)
-    XCTAssertEqual(valueType, "foo bar")
-    verify(inoutProtocol.parameterizedMethod(object: any())).wasCalled()
-  }
-  func testInoutParameter_uppercasesString() {
-    given(inoutProtocol.parameterizedMethod(object: any())) ~> { $0 = $0.uppercased() }
-    var valueType = "foo bar"
-    inoutProtocolInstance.parameterizedMethod(object: &valueType)
-    XCTAssertEqual(valueType, "FOO BAR")
-    verify(inoutProtocol.parameterizedMethod(object: any())).wasCalled()
   }
   
   // MARK: Chained stubbing
